requirements.txt
duckduckgo-search>=6.0.0
python-whois>=0.8.0
thefuzz>=0.20.0
python-Levenshtein>=0.25.0
fpdf2>=2.7.0
requests>=2.31.0
tldextract>=5.1.2
